register router working
