+++ api/src/routes/user.ts
@@ -1,7 +1,6 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { Router, Request, Response, NextFunction } from 'express';
-import { issueJWT } from '../lib/utils';
-import User from '../models/User';
+import register from "../controllers/user"
 
 const router = Router();
 
@@ -17,26 +16,6 @@ router.post(
   (_req: Request, _res: Response, _next: NextFunction) => {}
 );
 
-// TODO
-router.post(
-  '/register',
-  async (req: Request, res: Response, _next: NextFunction) => {
-    try {
-      const newUser = new User();
-      const password = await newUser.bcryptPassword(req.body.password);
-      const user = await newUser.registerUser(req.body.user_name, password);
-      const jwt = issueJWT(user);
-
-      res.json({
-        success: true,
-        user,
-        token: jwt.token,
-        expiresIn: jwt.expires
-      });
-    } catch (error) {
-      res.json({ success: false, error });
-    }
-  }
-);
+router.post('/register', register);
 
 export default router;
